requirements.txt
flask==2.3.3
opencv-python==4.8.1.78
pillow==10.0.1
requests==2.31.0
numpy==1.24.3